// code/frontend/frp-fe/src/api/models/AccExternalMappingDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AccExternalMappingDto = {
    id?: number;
    kind?: 'ACCOUNT' | 'CATEGORY';
    externalId?: string;
    externalName?: string;
    currencyCode?: string;
    accountId?: number;
    ignored?: boolean;
};